satochip: remove reference to crypto.py in SecureChannel.py
Use cryptolib/cryptography.py: aes_encrypt, aes_decrypt, append_PKCS7_padding, strip_PKCS7_padding instead
