+++ cryptolib/cryptography.py
@@ -169,6 +169,20 @@ def unpad_data(data):
         raise Exception("Bad padding in received data")
     return data[:i]
 
+def append_PKCS7_padding(data: bytes) -> bytes:
+    padlen = 16 - (len(data) % 16)
+    return data + bytes([padlen]) * padlen
+
+def strip_PKCS7_padding(data: bytes) -> bytes:
+    if len(data) % 16 != 0 or len(data) == 0:
+        raise Exception("invalid padding length")
+    padlen = data[-1]
+    if not (0 < padlen <= 16):
+        raise Exception("invalid padding byte (out of range)")
+    for i in data[-padlen:]:
+        if i != padlen:
+            raise Exception("invalid padding byte (inconsistent)")
+    return data[0:-padlen]
 
 def gen_iv():
     """Randomly generates 16 bytes data."""
